+++ src/main/java/ru/mark/springwebapp/service/UserService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserService {
     List<User> index();
-    User show(int id);
+    User show(long id);
     void save(User user);
-    void update(int id, User newUser);
-    void delete(int id);
+    void update(long id, User newUser);
+    void delete(long id);
 }
